sprint_510_2025

+++ system/module_page/backend/patterns/Form.php
@@ -182,18 +182,27 @@ namespace effcore {
   # ──────────────────────────────────────────────────────────────────────────────
 
   protected function validation_cache_select() {
-    $cache = (new instance('cache_validation', ['id' => $this->validation_id]))->select();
-    return $cache ? unserialize($cache->data) : [];
+    $storage = storage::get(entity::get('cache_validation')->storage_name);
+    if ($storage->is_available()) {
+      $cache = (new instance('cache_validation', ['id' => $this->validation_id]))->select();
+      return $cache ? unserialize($cache->data) : [];
+    }
   }
 
   protected function validation_cache_update($cache) {
-    $instance = new instance('cache_validation', ['id' => $this->validation_id]);
-    if ($instance->select()) {$instance->data = serialize($cache); return $instance->update();}
-    else                     {$instance->data = serialize($cache); return $instance->insert();}
+    $storage = storage::get(entity::get('cache_validation')->storage_name);
+    if ($storage->is_available()) {
+      $instance = new instance('cache_validation', ['id' => $this->validation_id]);
+      if ($instance->select()) {$instance->data = serialize($cache); return $instance->update();}
+      else                     {$instance->data = serialize($cache); return $instance->insert();}
+    }
   }
 
   protected function validation_cache_delete() {
-    return (new instance('cache_validation', ['id' => $this->validation_id]))->delete();
+    $storage = storage::get(entity::get('cache_validation')->storage_name);
+    if ($storage->is_available()) {
+      return (new instance('cache_validation', ['id' => $this->validation_id]))->delete();
+    }
   }
 
   function validation_cache_date_get($format = 'Y-m-d') {
